# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
#	src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/UserService.java

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -11,8 +11,10 @@ import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.utility.CheckForId;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -20,55 +22,61 @@ public class FilmService {
 
     @Qualifier("filmDbStorage")
     private final FilmStorage filmStorage;
+
+    @Qualifier("userDbStorage")
+    private final UserStorage userStorage;
     private final LikesDao likesDao;
-
     private final FilmDirectorDao filmDirectorDao;
-
     private final DirectorDao directorDao;
-
     private final FilmGenreDao filmGenreDao;
 
     @Autowired
     public FilmService(@Qualifier("filmDbStorage") FilmStorage inMemoryFilmStorage,
-                       LikesDao likesDao, FilmDirectorDao filmDirectorDao,
+                       LikesDao likesDao,
+                       FilmDirectorDao filmDirectorDao,
                        DirectorDao directorDao,
-                       FilmGenreDao filmGenreDao) {
+                       FilmGenreDao filmGenreDao,
+                       @Qualifier("userDbStorage") UserStorage userStorage) {
         this.filmStorage = inMemoryFilmStorage;
         this.likesDao = likesDao;
         this.filmDirectorDao = filmDirectorDao;
         this.directorDao = directorDao;
         this.filmGenreDao = filmGenreDao;
+        this.userStorage = userStorage;
     }
 
     private static final LocalDate DATE = LocalDate.of(1895, 12, 28);
 
     public Film create(Film film)
             throws AlreadyExistException, NotBurnYetException, IllegalLoginException {
-        if (film.getReleaseDate().isBefore(DATE)) {
+        if (film.getReleaseDate().isBefore(DATE))
             throw new IllegalArgumentException("Выбрана ложная дата релиза");
-        }
         if (!filmStorage.contains(film)) {
             log.info("Данный фильм добавлен");
             return filmStorage.saveFilm(film);
         } else {
-            log.error("Данный фильм уже добавлен");
+            log.warn("Данный фильм уже добавлен");
             throw new AlreadyExistException("Данный фильм уже добавлен");
         }
     }
 
     public Film update(Film film) throws NegativeIdException {
         CheckForId.idCheck(film.getId());
-        if (film.getReleaseDate().isBefore(DATE)) {
+        if (film.getReleaseDate().isBefore(DATE))
             throw new IllegalArgumentException("Выбрана ложная дата релиза");
+        filmStorage.update(film);
+        log.info("Данные о фильме добавлены или обновлены");
+        if (film.getDirectors() != null) {
+            filmDirectorDao.updateFilmDirector(film);
         }
-        return filmStorage.update(film);
+        return film;
     }
 
     public List<Film> get() {
 //        log.info("Текущее количество добавленных фильмов: {}", filmStorage.getSize());
         List<Film> films = filmStorage.findAll();
 
-        for(Film film: films) {
+        for (Film film : films) {
             long id = film.getId();
             film.setGenres(filmGenreDao.getFilmGenreById(id));
             filmDirectorDao.setFilmDirector(film);
@@ -80,7 +88,7 @@ public class FilmService {
     public Film getFilmById(Long id) throws NegativeIdException {
         CheckForId.idCheck(id);
         if (!filmStorage.contains(id)) {
-            log.error("Данный пользователь не существует");
+            log.warn("Данный пользователь не существует");
             throw new InvalidIdInPathException("Данный пользователь не существует");
         }
         log.info("Заданный пользователь успешно возвращен");
@@ -90,19 +98,39 @@ public class FilmService {
     public void putLikeToFilm(Long film_id, Long favId) throws NegativeIdException {
         CheckForId.idCheck(film_id);
         CheckForId.idCheck(favId);
-        likesDao.putLike(film_id, favId);
-        log.info("+1 лайк");
+        if (filmStorage.contains(film_id) && userStorage.contains(favId)) {
+            likesDao.putLike(film_id, favId);
+            log.info("+1 лайк");
+        } else
+            throw new InvalidIdInPathException("Ошибка один из пользователей не существует");
     }
 
     public void deleteLikeToFilm(Long film_id, Long hateId) throws NegativeIdException {
         CheckForId.idCheckEquals(film_id, hateId);
-        likesDao.deleteLike(film_id, hateId);
-        log.info("-1 лайк");
+        if (filmStorage.contains(film_id) && userStorage.contains(hateId)) {
+            likesDao.deleteLike(film_id, hateId);
+            log.info("-1 лайк");
+        }
     }
 
     public List<Film> getPopularFilms(Long count) throws NegativeIdException {
         log.info("Вывод рейтинга фильмов по количеству лайков");
         return filmStorage.getPopularFilms(count);
+    }
+
+
+    public void deleteFilm(Long film_id) throws NegativeIdException {
+        CheckForId.idCheck(film_id);
+        filmStorage.deleteFilm(film_id);
+        log.info("-1 фильм");
+    }
+
+
+    public List<Film> getCommonFilms(long userId, long friendId) {
+        if (userStorage.contains(userId) & userStorage.contains(friendId)) {
+            log.info("Cписок фильмов, отсортированных по популярности.");
+            return filmStorage.getCommonFilms(userId, friendId);
+        } else throw new InvalidIdInPathException("Ошибка один из пользователей не существует");
     }
 
 
@@ -139,5 +167,22 @@ public class FilmService {
             return filmStorage.searchFilmsByDirectorOrTitle(query);
         else
             throw new InvalidIdInPathException("Передан некорректный параметр запроса");
+    }
+
+    public List<Film> getPopularFilmsOrderByGenreYear(Optional<Long> genreId, Optional<Integer> year, long count) {
+        if (genreId.isPresent() && year.isPresent()) {
+            log.info("Вывод фильмов, по жанру и году.");
+            return filmStorage.getPopularFilmsOrderByGenreYear(genreId, year, count);
+        }
+        if (year.isPresent()) {
+            log.info("Вывод рейтинга фильмов по году");
+            return filmStorage.getPopularFilmsOrderByYear(year, count);
+        }
+        if (genreId.isPresent()) {
+            log.info("Вывод рейтинга фильмов по жанру");
+            return filmStorage.getPopularFilmsOrderByGenre(genreId, count);
+        }
+        log.info("Вывод рейтинга фильмов по количеству лайков");
+        return filmStorage.getPopularFilms(count);
     }
 }
